+++ randovania/games/prime1/exporter/patch_data_factory.py
@@ -224,6 +224,22 @@ def _random_factor(rng: Random, min: float, max: float, target: float):
 
 
 def _pick_random_point_in_aabb(rng: Random, aabb: list, room_name: str):
+    if room_name == "Artifact Temple":
+        center = [-373, 47, -30]
+        scale = [65, 50, 16]
+        return [rng.uniform(center[i] - scale[i] / 2, center[i] + scale[i] / 2) for i in range(3)]
+
+    if room_name == "Burn Dome":
+        if bool(rng.getrandbits(1)):
+            # Main Room
+            center = [577.5, -13.1, 34.9]
+            scale = [25, 22.4, 7.4]
+        else:
+            # Missile Room
+            center = [588.8, 39.5, 33.9]
+            scale = [9, 8, 5]
+        return [rng.uniform(center[i] - scale[i] / 2, center[i] + scale[i] / 2) for i in range(3)]
+
     # return a quasi-random point within the provided aabb, but bias towards being closer to in-bounds
     offset_xy = 0.0
     offset_max_z = 0.0
@@ -236,7 +252,6 @@ def _pick_random_point_in_aabb(rng: Random, aabb: list, room_name: str):
         "Triclops Pit",
         "Elite Quarters",
         "Quarantine Cave",
-        "Burn Dome",
         "Research Lab Hydra",
         "Research Lab Aether",
     ]
@@ -244,7 +259,6 @@ def _pick_random_point_in_aabb(rng: Random, aabb: list, room_name: str):
     if room_name in ROOMS_THAT_NEED_HELP:
         offset_xy = 0.1
         offset_max_z = -0.3
-
     x_factor = _random_factor(rng, 0.15 + offset_xy, 0.85 - offset_xy, 0.5)
     y_factor = _random_factor(rng, 0.15 + offset_xy, 0.85 - offset_xy, 0.5)
     z_factor = _random_factor(rng, 0.1, 0.8 + offset_max_z, 0.35)
@@ -706,13 +720,18 @@ class PrimePatchDataFactory(PatchDataFactory):
                     if self.configuration.shuffle_item_pos or node.extra.get("position_required"):
                         aabb = area.extra["aabb"]
                         pickup["position"] = _pick_random_point_in_aabb(self.rng, aabb, area.name)
-
-                        if node.extra.get("position_required"):
-                            # Scan this item through walls
-                            assert self.configuration.items_every_room
-                            pickup["jumboScan"] = True
+                        pickup["jumboScan"] = True
 
                     level_data[region.name]["rooms"][area.name]["pickups"].append(pickup)
+
+        if self.configuration.shuffle_item_pos:
+            # Allow temple cutscene without collecting item
+            level_data["Tallon Overworld"]["rooms"]["Artifact Temple"]["triggers"] = [
+                {
+                    "id": 0x00100470,
+                    "active": True,
+                }
+            ]
 
         # serialize room modifications
         if self.configuration.superheated_probability != 0: